requirements.txt
chromadb
sounddevice
scipy
stripe
elevenlabs
langgraph
langchain-core
langchain-community
langchain-openai
trustcall
passlib
streamlit
PyAudio
SpeechRecognition[whisper-local]
transformers==4.49.0
qrcode
Pillow
streamlit_autorefresh
nest_asyncio
pyppeteer